Refactor: extract traits code.
