+++ src/formats/pro/parse/type/item/type/armor.rs
@@ -1,4 +1,5 @@
 use super::super::super::*;
+use super::super::super::traits::*;
 
 pub(crate) fn instance<S: Read>(source: &mut S) -> Result<object::item::armor::Instance, errors::Error> {
     let mut ac_bytes = [0u8; 4];
