+++ src/formats/pro/parse/type/item/type/misc.rs
@@ -1,4 +1,5 @@
 use super::super::super::*;
+use super::super::super::traits::*;
 
 pub(crate) fn instance<S: Read>(source: &mut S) -> Result<object::item::misc::Instance, errors::Error> {
     let mut item_pid_bytes = [0u8; 4];
